Early-out ipv6 checks when running on App Engine
